% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_ChenWat.R
\name{ChenWat}
\alias{ChenWat}
\title{Calculate and Plot Chen & Watanabe 1989 Model Mortality Estimates}
\usage{
ChenWat(VB_params, age_seq = NULL, plot = TRUE)
}
\arguments{
\item{VB_params}{A list containing Von Bertalanffy parameters:
- Linf: Asymptotic length.
- K: Growth coefficient.
- t0: Theoretical age at length zero.
- tmax: Maximum age.}

\item{age_seq}{Optional; a numeric vector of ages at which to estimate mortality.
Defaults to a sequence from 0 to tmax.}

\item{plot}{Logical; indicates whether to return a ggplot object of the mortality estimates.
Defaults to TRUE.}
}
\value{
Depending on the plot parameter, either a data frame or a ggplot object.
}
\description{
This function implements the Chen & Watanabe 1989 method for estimating natural mortality.
It automatically generates an age sequence based on the Von Bertalanffy parameters provided,
and can optionally return a plot of the mortality estimates.
}
\examples{
library(ggplot2)

# Define parameters using VB_params
VB_params <- VB_params(45, 0.2, -0.5, 7, 0.005, 2.98, 13, 4)

# Estimate Natural mortality using Chen & Watanabe 1989 Model and plot the results
results_ChenWat <- ChenWat(VB_params, seq(0, VB_params$tmax, by = 1), TRUE)
results_ChenWat

# To get the data frame without plotting for further analysis:
table_ChenWat <- ChenWat(VB_params, seq(0, VB_params$tmax, by = 1), FALSE)
# Display the results as a data frame
print(table_ChenWat)
# Plot the results manually
ggplot(table_ChenWat, aes(x = age, y = M_age)) +
  geom_line() + geom_point() +
  labs(title = "Chen & Watanabe 1989 Natural Mortality Estimates",
       x = "Age", y = bquote("Natural Mortality Rate"~(year^-1))) +
  theme_classic()

}
