% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoenig.R
\name{hoenig}
\alias{hoenig}
\title{Estimate natural mortality from longevity using Hoenig (1983)}
\usage{
hoenig(t_max)
}
\arguments{
\item{t_max}{maximum age of the fish or longevity}
}
\value{
natural mortality coefficient M
}
\description{
Estimate natural mortality from longevity using Hoenig (1983)
}
\examples{
\dontrun{
hoenig(t_max = 6)
}
}
