% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_ProdBiom.R
\name{ProdBiom}
\alias{ProdBiom}
\title{Calculate and Plot Abella et al. (1998) Model Mortality Estimates}
\usage{
ProdBiom(
  VB_params,
  N0 = 1e+05,
  initial_Ma = 0.2,
  initial_B = 0.4,
  max_iterations = 50,
  tolerance = 0.01,
  plot = TRUE
)
}
\arguments{
\item{VB_params}{A list from `VB_params()` containing growth and biological parameters.}

\item{N0}{Initial population size (default = 100000).}

\item{initial_Ma}{Initial guess for parameter Ma.}

\item{initial_B}{Initial guess for parameter B.}

\item{max_iterations}{Maximum optimization iterations.}

\item{tolerance}{Optimization convergence threshold.}

\item{plot}{Logical; if TRUE returns a ggplot object; else a data.frame.}
}
\value{
Either a `data.frame` or a `ggplot` object.
}
\description{
Estimates age-specific natural mortality using the Abella et al. (1998) method.
}
